structure update and Cart CategoryBrand updates

+++ cart-service/src/main/java/smr/shop/cart/service/dto/response/CartResponse.java
@@ -3,6 +3,7 @@ package smr.shop.cart.service.dto.response;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -21,7 +22,7 @@ public class CartResponse {
 
     Double price;
 
-    Double discountPrice;
+    BigDecimal discountPrice;
 
     List<CartItemResponse> items;
 }
